Create All distinct pairs with given sum.cpp
Given an array arr[] and an integer target. You have to find all distinct pairs in array arr[] which sums up to given target.

Note: 

(a, b) and (b, a) are considered the same. Also, the same numbers at different indices are considered same.
Return the list of pairs sorted lexicographically by the first element, and then by the second element if necessary.