Create Balance with respect to an array
Find whether a given number x is balanced with respect to a given array arr. Sort the array first in non-decreasing order. Check if the absolute difference between the floor of x and x is equal to the absolute difference between the ceil of x and x.
Note: The ceiling of x is the smallest element in the array greater than or equal to x, and the floor of x is the greatest element smaller than or equal to x. 

Examples:

Input: arr[] = [1, 2, 8, 10, 10, 12, 19], x = 5 
Output: true
Explanation: Here 2 is the floor value and 8 is the ceil value and (5 - 2) = (8 - 5).  
Input: arr[] = [1, 2, 5, 7, 8, 11, 12, 15], x = 9
Output: false
Explanation: Here 8 is the floor value and 11 is the ceil value and (9 - 8) != (11 - 9).  
Expected Time Complexity: O(nlogn)
Expected Auxiliary Space: O(1)